Moving core model in api app.
